+++ src/components/ProjectsSection/index.tsx
@@ -1,1 +1,2 @@
+export { ProjectsList } from './projectsList';
 export { Projects } from './projects';
